# docs/swagger.yaml
basePath: /
definitions:
...
  db.YoutubeDownload:
    properties:
      youtube_urls:
        items:
          type: string
        type: array
    type: object
  feeds.Destination:
    properties:
      address:
        type: string
      customKey:
        type: string
      customValue:
        type: string
      split: {}
      type:
        type: string
    type: object
  feeds.Episode:
    properties:
      datePublished:
        type: integer
      description:
        type: string
      enclosureLength:
        type: integer
      enclosureType:
        type: string
      enclosureUrl:
        type: string
      feedId:
        type: integer
      feedUrl:
        description: for search
        type: string
      id:
        type: integer
      image:
        type: string
      link:
        type: string
      title:
        type: string
    type: object
  feeds.Feed:
    properties:
      author:
        type: string
      contentType:
        type: string
      datePublished:
        type: integer
      dateUpdated:
        type: integer
      description:
        type: string
      feedType:
        description: podcast, video, blog
        type: integer
      generator:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      itemId:
        type: string
      items:
        items:
          $ref: '#/definitions/feeds.Item'
        type: array
      language:
        type: string
      link:
        type: string
      ownerUrl:
        type: string
      title:
        type: string
      url:
        type: string
      value:
        $ref: '#/definitions/feeds.Value'
    type: object
  feeds.Item:
    properties:
      author:
        type: string
      datePublished:
        type: integer
      dateUpdated:
        type: integer
      description:
        type: string
      duration:
        type: integer
      enclosureType:
        type: string
      enclosureUrl:
        type: string
      feedId:
        description: for search
        type: string
      feedType:
        type: integer
      id:
        type: string
      imageUrl:
        type: string
      link:
        type: string
      thumbnailUrl:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  feeds.Model:
    properties:
      suggested:
        type: string
      type:
        type: string
    type: object
  feeds.Podcast:
    properties:
      author:
        type: string
      contentType:
        type: string
      description:
        type: string
      episodes:
        items:
          $ref: '#/definitions/feeds.Episode'
        type: array
      generator:
        type: string
      id:
        type: integer
      image:
        type: string
      language:
        type: string
      lastUpdateTime:
        type: integer
      link:
        type: string
      title:
        type: string
      url:
        type: string
      value:
        $ref: '#/definitions/feeds.Value'
    type: object
  feeds.Value:
    properties:
      destinations:
        items:
          $ref: '#/definitions/feeds.Destination'
        type: array
      model:
        $ref: '#/definitions/feeds.Model'
    type: object
...
host: localhost:5002
info:
  contact: {}
  description: This is the API documentation for Sphinx Tribes.
  title: Sphinx Tribes API
  version: "1.0"
paths:
...
  /feed:
    get:
      consumes:
      - application/json
      description: Get a generic feed by URL
      parameters:
      - description: Feed URL
        in: query
        name: url
        required: true
        type: string
      - description: Tribe UUID
        in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/feeds.Feed'
      summary: Get Generic Feed
      tags:
      - Feeds
  /feed/download:
    post:
      consumes:
      - application/json
      description: Download a Youtube feed
      parameters:
      - description: Youtube Download
        in: body
        name: youtube_download
        required: true
        schema:
          $ref: '#/definitions/db.YoutubeDownload'
      produces:
      - application/json
      responses:
        "200":
          description: Youtube download processed successfully
          schema:
            type: string
      summary: Download Youtube Feed
      tags:
      - Feeds
...
  /podcast:
    get:
      consumes:
      - application/json
      description: Get a podcast by URL or ID
      parameters:
      - description: Feed URL
        in: query
        name: url
        type: string
      - description: Feed ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/feeds.Podcast'
      summary: Get Podcast
      tags:
      - Feeds
...
  /search_podcast_episodes:
    get:
      consumes:
      - application/json
      description: Search for podcast episodes by query
      parameters:
      - description: Search Query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feeds.Item'
            type: array
      summary: Search Podcast Episodes
      tags:
      - Feeds
  /search_podcasts:
    get:
      consumes:
      - application/json
      description: Search for podcasts by query
      parameters:
      - description: Search Query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feeds.Feed'
            type: array
      summary: Search Podcasts
      tags:
      - Feeds
  /search_youtube:
    get:
      consumes:
      - application/json
      description: Search for Youtube videos by query
      parameters:
      - description: Search Query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feeds.Feed'
            type: array
      summary: Search Youtube
      tags:
      - Feeds
  /search_youtube_videos:
    get:
      consumes:
      - application/json
      description: Search for Youtube videos by query
      parameters:
      - description: Search Query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/feeds.Item'
              type: array
            type: array
      summary: Search Youtube Videos
      tags:
      - Feeds
...
  /youtube_videos:
    get:
      consumes:
      - application/json
      description: Get Youtube videos for a specific channel
      parameters:
      - description: Channel ID
        in: query
        name: channelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/feeds.Item'
              type: array
            type: array
      summary: Get Youtube Videos for Channel
      tags:
      - Feeds
swagger: "2.0"